Move edit functions to components
